FIX: tensor to float

+++ src/utils.py
@@ -80,7 +80,7 @@ class GetResps(object):
                     if mask[cx:cx+wsize[0], cy:cy+wsize[1]].sum() != wsize[0] * wsize[1]:
                         resps[:, i, j] = min_feat
                     
-        return self.to_torch(resps)
+        return self.to_torch(resps).float()
 
 
 def init_weights(model, ref_model):
